Moved legacy map layers to be more clear
