+++ src/_legacy/map-dev/column-inspector/modal-panel.ts
