+++ src/_legacy/map-sources/components/app.ts
@@ -1,5 +1,5 @@
 import { useEffect } from "react";
-import { useBurwellActions } from "~/burwell-sources/app-state";
+import { useBurwellActions } from "~/_legacy/map-sources/app-state";
 import hyper from "@macrostrat/hyper";
 import IndexMapContainer from "./map";
 import { InfoDrawer } from "./info-drawer";
